Add a grammar to parse a specific JSON string
The tool is used in connection with ABS REST API.
The REST API in ABS returns the result of HTTP requests
in a specialized JSON format which is not generic.

The newly added 'metrics.py' uses Python 'modgrammar'
to define a proper grammar of the JSON string to allow
parsing the HTTP response on the fly without the need
for a custom string parsing/implementation.

Signed-off-by: Behrooz Nobakht <nobeh5@gmail.com>